% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bats.R
\docType{data}
\name{bats}
\alias{bats}
\title{Percent Body Fat of Little Brown Bats from Aeolus Cave}
\format{
A data frame with 159 rows and 4 variables:
\describe{
  \item{sex}{sex of the sampled bat, with \code{F} for female and \code{M} for male.}
  \item{percentfat}{percent body fat.}
  \item{year}{year the bat was sampled.}
  \item{days}{hibernation time, defined as the number of days since the fall equinox.}
}
}
\source{
\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.sh487nh}
}
\usage{
data(bats)
}
\description{
This dataset contains information on the percent body fat of little brown bats (\emph{Myotis lucifugus})
    sampled in Aeolus Cave, East Dorset, Vermont, located in the northeastern United States.
}
\details{
This subset of data was collected by Cheng et al. (2019) specifically from Aeolus Cave, one of
    five regions studied in the United States. The dataset provides insight into the fat stores
    of bats, which is crucial for understanding their persistence in the face of white-nose syndrome.
}
\references{
Cheng TL, Gerson A, Moore MS, et al. (2019). Higher fat stores contribute to persistence of little brown bat populations with white-nose syndrome. \emph{Journal of Animal Ecology}. \bold{88}, 591--600
}
\keyword{datasets}
