% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_diagnostic.R
\name{envelope}
\alias{envelope}
\alias{print.envelope}
\alias{plot.envelope}
\title{Normal Probability Plot with Simulated Envelope}
\usage{
envelope(
  object,
  type = c("quantile", "pearson", "response", "deviance"),
  nsim = 99,
  level = 0.95,
  plot = TRUE,
  progressBar = TRUE,
  ...
)

\method{print}{envelope}(x, ...)

\method{plot}{envelope}(
  x,
  type = c("quantile", "pearson", "response", "deviance"),
  level = 0.95,
  ...
)
}
\arguments{
\item{object, x}{an \code{"ugrpl"} object.}

\item{type}{character; specifies which residual should be produced in the
envelope plot. The available options are \code{"quantile"} (default),
\code{"pearson"}, \code{"response"} (raw residuals, y - mu), and \code{"deviance"}.}

\item{nsim}{the number of replicates. The default is \code{nsim = 99}.}

\item{level}{confidence level for the construction of the confidence bands.}

\item{plot}{logical. If \code{TRUE} (the default) the graph is plotted.}

\item{progressBar}{logical; if \code{TRUE}, a progress bar is displayed giving
the progress of making the graph. It can slow down the function
considerably in applications with a large sample size.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{envelope} returns an \code{"envelope"} object witch consists of
    a list with the following components:
\describe{
  \item{residuals}{a list with the quantile, pearson, raw, and deviance residuals resulting from
    the fit of the unit gamma regression with parametric link function.}
  \item{simulation}{a list whose components are matrices containing the
      ordered quantile, pearson, raw, and deviance residuals of the simulation for the
      envelope plot.}

 The method \code{plot} makes the envelope plot.
 }
}
\description{
Provides the normal probability plot with simulated envelope of the residuals
    resulting from the fit of the unit gamma regression with parametric link function.
}
