% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_diagnostic.R
\name{influence}
\alias{influence}
\title{Title}
\usage{
influence(
  object,
  scheme = c("case-weight", "mean", "dispersion", "both"),
  covariate,
  plot = TRUE,
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  ...
)
}
\arguments{
\item{object}{an \code{"ugrpl"} object.}

\item{scheme}{character; it specifies the perturbation scheme. Currently, the following
perturbation schemes are available: case-weight perturbation (\code{"case-weight"});
perturbation of the mean covariates (\code{"mean"}); perturbation of the dispersion
covariates (\code{"dispersion"}); and simultaneous perturbation of the mean and dispersion
covariates (\code{"both"}). If the mean and the dispersion submodels share the same
covariate, then it is not possible to use \code{scheme = "mean"} or \code{scheme = "dispersion"}.}

\item{covariate}{character; it specifies the name of the continuous variable used in the covariate
perturbation scheme. If \code{scheme = "both"}, then it must be a vector of size 2, where
the first element is the name of the perturbed covariate associated with the mean and the
second element is the name of the perturbed covariate associated with the dispersion.
If \code{scheme = "mean"} or \code{scheme = "dispersion"}, then the mean and dispersion
submodels cannot share the same perturbed covariate.}

\item{plot}{logical. If \code{TRUE} (the default) the graph is plotted.}

\item{ask}{logical. If \code{TRUE}, the user is asked before each plot.}

\item{...}{further arguments passed \link{plot}.}
}
\value{
A list containing the following components:
\describe{
  \item{dmax}{local influence under the selected perturbation scheme.}
  \item{total.LI}{total local influence.}
  \item{scheme}{name of the considered perturbation scheme}
  \item{leverage}{the diagonal elements of the generalized leverage matrix.}
 }
}
\description{
Title
}
