% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_plink_choice.R
\name{plink_choice-methods}
\alias{plink_choice-methods}
\alias{print.plink_choice}
\title{Methods for \code{"plink_choice"} objects}
\usage{
\method{print}{plink_choice}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"plink_choice"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Methods for \code{"plink_choice"} objects
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
