% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_plink_choice.R
\name{plink_choice}
\alias{plink_choice}
\title{Choosing the Link Function in the Unit Gamma Regression}
\usage{
plink_choice(
  object,
  link,
  sigma.link,
  criterion = c("aic", "bic", "upsilon"),
  control = ug_control(...),
  ...
)
}
\arguments{
\item{object}{a "\code{ugrpl}" object.}

\item{link, sigma.link}{a vector of characters specifying which link functions should be
considered for the mean and dispersion parameter submodels, respectively. By default,
possible combinations between the link functions \code{"logit"}, \code{"probit"},
\code{"plogit"}, \code{"pprobit"}, \code{"rplogit"}, \code{"rpprobit"}, \code{"aordaz"},
and \code{"raordaz"} are considered. See \code{\link{make.plink}}.}

\item{criterion}{criteria for choosing the "best" combination. AIC, BIC, or Upsilon statistics
can be used.}

\item{control}{a list of control parameters passed as arguments for the \code{optim} function
specified via \code{\link{ug_control}}.}

\item{...}{arguments passed to \code{\link{ug_control}}.}
}
\value{
The \code{plink.choice} function returns a list where each component is a unit gamma
    regression fit with a combination of the links provided by the \code{link} and \code{sigma.link}
    arguments.
}
\description{
Choosing the Link Function in the Unit Gamma Regression
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
