% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_diagnostic.R
\name{plot.ugrpl}
\alias{plot.ugrpl}
\title{Diagnostic plots for the Unit Gamma regression with parametric link
 function}
\usage{
\method{plot}{ugrpl}(
  x,
  which = 1:4,
  type = c("quantile", "pearson", "response", "deviance"),
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  ...
)
}
\arguments{
\item{x}{an object of class \code{"ugrpl"}.}

\item{which}{numeric. If a subset of the plots is required, specify a subset
of the numbers \\code{1:8}. See details below.}

\item{type}{character; it specifies which residual should be produced
in the summary plot. The available arguments are \code{"quantile"} (default),
 \code{"pearson"}, \code{"response"} (raw residuals, y - mu), and \code{"deviance"}.}

\item{ask}{logical. If \code{TRUE}, the user is asked before each plot.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Diagnostic plots for the Unit Gamma regression with parametric link
 function
}
\details{
The \code{which} argument can be used to select a subset of currently eight supported
    types of displays. The displays are:

\describe{
  \item{\code{which = 1}:}{residuals vs fitted values}
  \item{\code{which = 2}:}{residuals vs indices of observations}
  \item{\code{which = 3}:}{normal probability plot of the residuals. Recall that only the
      quantile residuals have a standard normal distribution.}
  \item{\code{which = 4}:}{local influence versus indices of observations under the case-weights
      perturbation scheme.}
  \item{\code{which = 5}:}{generalized leverage versus indices of observations.}
  \item{\code{which = 6}:}{response vs fitted values.}
  \item{\code{which = 7}:}{autocorrelation plot of the residuals.}
  \item{\code{which = 8}:}{partial autocorrelation plot of the residuals.}
 }
}
