% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_methods.R
\name{predict.ugrpl}
\alias{predict.ugrpl}
\title{Predict Method for Unit Gamma Regression with Parametric Link Functions}
\usage{
\method{predict}{ugrpl}(
  object,
  newdata = NULL,
  type = c("response", "link", "dispersion", "variance", "quantile"),
  na.action = stats::na.pass,
  at = 0.5,
  ...
)
}
\arguments{
\item{object}{a \code{'ugrpl'} object.}

\item{newdata}{optionally, a data frame in which to look for variables
with which to predict. If omitted, the fitted linear predictors are
used.}

\item{type}{the type of prediction required. The default is on the scale of
the response variable \code{("response")}, that is, the fitted values
(fitted means). The alternative \code{"link"} is on the scale of
the linear predictors. The
specification \code{"dispersion"} provides the fitted dispersion parameters,
while \code{"variance"} provides the fitted variances. Finally,
the option \code{"quantile"} gives the fitted quantiles in the order
specified via \code{'at'}.}

\item{na.action}{function determining what should be done with missing
values in \code{newdata}. The default is to predict \code{NA}.}

\item{at}{the order of the quantile to be predicted if
\code{type = "quantile"}. The default is to predict the median,
that is, \code{at = 0.5}.}

\item{...}{arguments passed to or from other methods.}
}
\value{
A vector of predictions.
}
\description{
Obtains predictions from a fitted unit gamma regression object.
}
