% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_diagnostic.R
\name{reset}
\alias{reset}
\title{RESET-Type Goodness-of-Fit Test}
\usage{
reset(object)
}
\arguments{
\item{object}{an object of class \code{"ugrpl"}.}
}
\value{
A list with the following components:
\describe{
  \item{statistic:}{the value the likelohood ratio test statistic.}
  \item{p.value:}{the p-value for the test.}
 }
}
\description{
\code{reset} performs the RESET-type goodness-of-fit test to verify the correct specification
    of the unit gamma regression with parametric link function.
}
\details{
...
}
