% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_methods.R
\name{residuals.bergreg}
\alias{residuals.bergreg}
\alias{residuals.ugrpl}
\title{Extract Model Residuals for a Unit Gamma Regression with Parametric Link Functions}
\usage{
\method{residuals}{ugrpl}(object, type = c("quantile", "pearson", "response", "deviance"), ...)
}
\arguments{
\item{object}{a \code{'ugrpl'} object.}

\item{type}{character; it specifies which residual should be extracted.
The available arguments are \code{"quantile"} (default), \code{"pearson"},
\code{"response"} (raw residuals, y - mu), and \code{"deviance"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extract Model Residuals for a Unit Gamma Regression with Parametric Link Functions
}
