% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_maximum_likelihood_estimation.R
\name{ug_mle}
\alias{ug_mle}
\title{Title}
\usage{
ug_mle(
  y,
  X,
  Z = NULL,
  link = "aordaz",
  sigma.link = NULL,
  control = ug_control(...),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
