% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_methods.R
\name{ugrpl-methods}
\alias{ugrpl-methods}
\alias{model.frame.ugrpl}
\alias{model.matrix.ugrpl}
\alias{coef.ugrpl}
\alias{vcov.ugrpl}
\alias{logLik.ugrpl}
\alias{AIC.ugrpl}
\alias{print.ugrpl}
\alias{summary.ugrpl}
\alias{print.summary.ugrpl}
\title{Methods for 'ugrpl' objects}
\usage{
\method{model.frame}{ugrpl}(formula, ...)

\method{model.matrix}{ugrpl}(object, model = c("mean", "dispersion"), ...)

\method{coef}{ugrpl}(object, model = c("full", "mean", "dispersion"), ...)

\method{vcov}{ugrpl}(object, model = c("full", "mean", "dispersion", "lambda"), ...)

\method{logLik}{ugrpl}(object, ...)

\method{AIC}{ugrpl}(object, ..., k = 2)

\method{print}{ugrpl}(x, digits = getOption("digits"), ...)

\method{summary}{ugrpl}(object, ...)

\method{print}{summary.ugrpl}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{formula}{a model formula or terms object or an R object.}

\item{...}{further arguments passed to or from other methods.}

\item{model}{a character indicating which parameter coefficients are
required, parameters for the \code{"mean"} or for the
\code{"dispersion"} model. If \code{"full"} (default), a list with
coefficients for the \code{mean} and for the \code{dispersion}
model is returned.}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical AIC.}

\item{x, object}{an object of class \code{ugrpl}.}

\item{digits}{a non-null value for digits specifies the minimum number of significant digits to
be printed in values. The default, \code{getOption("digits")}.}
}
\description{
Methods for 'ugrpl' objects
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
