% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_diagnostic.R
\name{influence}
\alias{influence}
\title{Leverage and Local Influence for the Unit Gamma Regression}
\usage{
influence(
  object,
  scheme = c("case-weight", "mean", "dispersion", "both"),
  covariate,
  plot = TRUE,
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  ...
)
}
\arguments{
\item{object}{an \code{"ugrpl"} object.}

\item{scheme}{character; it specifies the perturbation scheme. Currently, the following
perturbation schemes are available: case-weight perturbation (\code{"case-weight"});
perturbation of the mean covariates (\code{"mean"}); perturbation of the dispersion
covariates (\code{"dispersion"}); and simultaneous perturbation of the mean and dispersion
covariates (\code{"both"}). If the mean and the dispersion submodels share the same
covariate, then it is not possible to use \code{scheme = "mean"} or \code{scheme = "dispersion"}.}

\item{covariate}{character; it specifies the name of the continuous variable used in the covariate
perturbation scheme. If \code{scheme = "both"}, then it must be a vector of size 2, where
the first element is the name of the perturbed covariate associated with the mean and the
second element is the name of the perturbed covariate associated with the dispersion.
If \code{scheme = "mean"} or \code{scheme = "dispersion"}, then the mean and dispersion
submodels cannot share the same perturbed covariate.}

\item{plot}{logical. If \code{TRUE}, the graphs of local influence and generalized leverage for each
sampled observation are presented.}

\item{ask}{logical. If \code{TRUE}, the user is asked before each plot.}

\item{...}{further arguments passed \link{plot}.}
}
\value{
A list containing the following components:
\describe{
  \item{dmax}{local influence under the selected perturbation scheme.}
  \item{total.LI}{total local influence.}
  \item{scheme}{name of the considered perturbation scheme}
  \item{leverage}{the diagonal elements of the generalized leverage matrix.}
 }
}
\description{
Leverage and Local Influence for the Unit Gamma Regression
}
\details{
Leverage points are observations that have a disproportionate weight in their fitted value.
    These points, in general, present a different behavior from the other points in relation
    to the values of the explanatory variables and can strongly influence the estimates of the
    regression coefficients. On the other hand, local influence analysis should be considered
    when there is an interest in investigating the sensitivity of postulated assumptions under
    small perturbations in the model or data.

The leverage measure considered is based on the work of Wei et al. (1998) and local influence
    perturbation schemes are commonly considered schemes, see, for instance Cook (1986).
}
\references{
Cook, R. D. (1986). Assessment of local influence. \emph{Journal of the Royal Statistical
    Society  B}, \bold{48}, 133--155.

Wei, B. C., Hu, Y. Q., & Fung, W. K. (1998). Generalized leverage and its applications.
    \emph{Scandinavian Journal of Statistics}, \bold{25}, 25--37.
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
